composer/backend/requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
eventlet==0.39.1
kafka-python==2.1.5
python-dotenv==1.0.1
python-socketio[client]==5.12.1  # Suporte ao cliente WebSocket
python-engineio==4.11.2
pymysql==1.1.0
redis==4.5.1 